Database only option in scripts

+++ scripts/reset_local_data_with_production_dump.sh
@@ -1,10 +1,29 @@
 #!/usr/bin/env bash
 
+function usage {
+	cat <<EOM
+Usage: $(basename "$0") [OPTION]...
+
+  -d          reset only database, not active storage blobs
+  -h          display help
+EOM
+
+	exit 2
+}
+
+DATABASE_ONLY=false
+
+while getopts 'd?h' c
+do
+  case $c in
+    d) DATABASE_ONLY=true ;;
+    h|?) usage ;;
+  esac
+done
+
 cd /tmp
-rm demagog_production_db.sql
+rm -f demagog_production_db.sql
 unzip -q demagog_production_db.zip
-rm -r demagog_production_active_storage_blobs
-unzip -q demagog_production_active_storage_blobs.zip
 cd -
 
 psql -U postgres -c 'REVOKE CONNECT ON DATABASE demagog_development FROM public;'
@@ -15,14 +34,21 @@ psql -U postgres -c 'DROP DATABASE demagog_development';
 psql -U postgres -c 'CREATE DATABASE demagog_development';
 psql -d demagog_development -U postgres -f /tmp/demagog_production_db.sql
 
-ACTIVE_STORAGE_BLOBS_DUMP_DIR=/tmp/demagog_production_active_storage_blobs
-LOCAL_STORAGE_DIR="`dirname $0`/../storage"
+if ! $DATABASE_ONLY; then
+  cd /tmp
+  rm -rf demagog_production_active_storage_blobs
+  unzip -q demagog_production_active_storage_blobs.zip
+  cd -
 
-rm -rf $LOCAL_STORAGE_DIR/*
+  ACTIVE_STORAGE_BLOBS_DUMP_DIR=/tmp/demagog_production_active_storage_blobs
+  LOCAL_STORAGE_DIR="`dirname $0`/../storage"
 
-for filename in `/bin/ls "$ACTIVE_STORAGE_BLOBS_DUMP_DIR" | xargs`
-do
-  dir=`echo $filename | cut -c 1-2`/`echo $filename | cut -c 3-4`
-  mkdir -p "$LOCAL_STORAGE_DIR/$dir"
-  mv "$ACTIVE_STORAGE_BLOBS_DUMP_DIR/$filename" "$LOCAL_STORAGE_DIR/$dir"
-done
+  rm -rf $LOCAL_STORAGE_DIR/*
+
+  for filename in `/bin/ls "$ACTIVE_STORAGE_BLOBS_DUMP_DIR" | xargs`
+  do
+    dir=`echo $filename | cut -c 1-2`/`echo $filename | cut -c 3-4`
+    mkdir -p "$LOCAL_STORAGE_DIR/$dir"
+    mv "$ACTIVE_STORAGE_BLOBS_DUMP_DIR/$filename" "$LOCAL_STORAGE_DIR/$dir"
+  done
+fi
